Merge tweaks

+++ pkg/client/service.go
@@ -528,16 +528,20 @@ func (c *Client) UpdateFile(updatedFile *svc.File) error {
 
 // remove a file in a specied directory. removes the file from the server too.
 func (c *Client) RemoveFile(file *svc.File) error {
+	// make sure this file is actually registered with the service
+	// before mucking around.
+	if !c.KnownItem(file.ClientPath) {
+		return fmt.Errorf("file %s not registered", file.Name)
+	}
 	// stop monitoring the file
 	c.Monitor.StopWatching(file.Path)
-
 	// we're implementing "soft" deletes here. if a user wants to
-	// actually Delete a file, we can implement another function for that later.
-	// remove from drive and database
+	// actually Delete a file, then users will pass the --delete flag
+	// to the remove command, which will fully delete the file.
 	if err := file.Copy(filepath.Join(c.RecycleBin, file.Name)); err != nil {
 		return fmt.Errorf("failed to copy file to recyle directory: %v", err)
 	}
-	// remove physical file from original location
+	// **remove physical file from original location**
 	if err := c.Drive.RemoveFile(file.DirID, file); err != nil {
 		return err
 	}
@@ -1010,62 +1014,6 @@ client-side DB. If there are any on the server that aren't on the client
 side, pull those files from the server and add to the local client service.
 */
 func (c *Client) RefreshFromServer() error { return nil }
-
-// discover populates the given root directory with the users file and
-// sub directories, updates the database as it does so, and returns
-// the the directory object when finished, or if there was an error.
-//
-// this should ideally be used for starting a new sfs service in a
-// users root directly that already has files and/or subdirectories.
-func (c *Client) DiscoverInRoot(root *svc.Directory) (*svc.Directory, error) {
-	// traverse users SFS file system and populate internal structures
-	root.Walk()
-
-	// send everything to the database
-	files := root.GetFiles()
-	c.log.Info(fmt.Sprintf("adding %d files...", len(files)))
-
-	if err := c.Db.AddFiles(files); err != nil {
-		return root, fmt.Errorf("failed to add file to database: %v", err)
-	}
-	for _, file := range files {
-		if err := c.WatchItem(file.Path); err != nil {
-			return root, err
-		}
-		if c.autoSync() {
-			if err := c.RegisterFile(file); err != nil {
-				return root, err
-			}
-		}
-	}
-
-	dirs := root.GetSubDirs()
-	c.log.Info(fmt.Sprintf("adding %d directories...", len(dirs)))
-
-	if err := c.Db.AddDirs(dirs); err != nil {
-		return root, fmt.Errorf("failed to add directory to database: %v", err)
-	}
-	for _, subDir := range dirs {
-		// if err := c.WatchItem(subDir.Path); err != nil {
-		// 	return root, err
-		// }
-		if c.autoSync() {
-			if err := c.RegisterDirectory(subDir); err != nil {
-				return root, err
-			}
-		}
-	}
-
-	// add root directory itself
-	c.log.Info("adding root...")
-	if err := c.Db.AddDir(root); err != nil {
-		return nil, fmt.Errorf("failed to add root to database: %v", err)
-	}
-	// if err := c.WatchItem(root.Path); err != nil {
-	// 	return nil, err
-	// }
-	return root, nil
-}
 
 // similar to DiscoverInRoot, but uses a specified directory path
 // and does not return a new directory object.
